change file structure and add search function
