+++ src/app/RecipeManagement/ManageRecipe/CreateRecipeForm/page.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { useForm } from "react-hook-form";
 import { motion } from "framer-motion";
 import axios from "axios";
-import { useRouter, useSearchParams } from "next/navigation";
 import toast from 'react-hot-toast';
+import { useRouter } from 'next/navigation';
 
 type RecipeFormData = {
   title: string;
@@ -23,44 +23,22 @@ type RecipeFormData = {
 
 const steps = ["Basic Info", "Details", "Nutrition & Image"];
 
-const UpdateRecipePage = () => {
-  const [step, setStep] = useState(0);
-  const [isClient, setIsClient] = useState(false); // To track client-side rendering
-  const [recipeData, setRecipeData] = useState<RecipeFormData | null>(null);
-  const { register, handleSubmit, formState: { errors }, reset } = useForm<RecipeFormData>();
+const CreateRecipePage = () => {
   const router = useRouter();
-  const searchParams = useSearchParams();
-  const recipeId = searchParams.get('id'); // Get the recipe ID from search params
-  
-  // Set client-side flag after the component is mounted
-  useEffect(() => {
-    setIsClient(true);
-  }, []);
-
-  // Fetch recipe data when the page loads (client-side)
-  useEffect(() => {
-    if (recipeId && isClient) {
-      console.log("Fetching recipe with ID:", recipeId); // Log the recipeId to ensure it's correct
-      const fetchRecipeData = async () => {
-        try {
-          const response = await axios.get(`${process.env.NEXT_PUBLIC_API_BASE_URL}/api/Recipe/${recipeId}`);
-          console.log("Fetched recipe data:", response.data); // Log the fetched recipe data
-          setRecipeData(response.data);
-          reset(response.data); // Populate the form with fetched data
-        } catch (error) {
-          console.error("Error fetching recipe data:", error);
-          alert("Failed to load recipe data");
-        }
-      };
-      fetchRecipeData();
-    }
-  }, [recipeId, isClient, reset]);
+  const [step, setStep] = useState(0);
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+    reset,
+  } = useForm<RecipeFormData>();
 
   const nextStep = () => setStep((prev) => Math.min(prev + 1, steps.length - 1));
   const prevStep = () => setStep((prev) => Math.max(prev - 1, 0));
 
   const onSubmit = async (data: RecipeFormData) => {
-    console.log("Form updated:", data); // Log the form data on submit
+    console.log("Form submitted:", data);
+
     try {
       const formData = new FormData();
       formData.append("title", data.title);
@@ -78,21 +56,23 @@ const UpdateRecipePage = () => {
         formData.append("image", data.imageFile[0]); // backend must accept this as IFormFile
       }
 
-      const response = await axios.put(`${process.env.NEXT_PUBLIC_API_BASE_URL}/api/Recipe/updateRecipe/${recipeId}`, formData);
-      console.log("Recipe updated successfully:", response.data);
-      toast.success("Recipe updated successfully!");
+      const response = await axios.post(`${process.env.NEXT_PUBLIC_API_BASE_URL}/api/Recipe/addRecipe1`, formData, {
+        headers: {
+          'Content-Type': 'multipart/form-data',
+        },
+      });
 
+      console.log("Recipe added successfully:", response.data);
+      toast.success("Recipe added successfully!");
       reset();
-      setStep(0); // Reset to the first step after submission
-
+      setStep(0);
       setTimeout(() => {
-        router.replace('/Pages/RecipeManagement/ManageRecipe/ManageRecipe'); // Navigate to the RecipeManage page
+        router.replace('/ManageRecipe/ManageRecipe');
       }, 500);
-      
     } catch (error: unknown) {
       if (axios.isAxiosError(error)) {
         console.error("Axios error:", error.response?.data || error.message);
-        toast.error("Failed to update recipe: " + (error.response?.data?.message || error.message));
+        toast.error("Failed to add recipe: " + (error.response?.data?.message || error.message));
       } else {
         console.error("Unexpected error:", error);
         toast.error("An unexpected error occurred.");
@@ -100,34 +80,30 @@ const UpdateRecipePage = () => {
     }
   };
 
-  if (!isClient) return null; // Avoid rendering anything before client-side hydration
-  if (!recipeData) return <p>Loading recipe data...</p>; // Show loading message until recipe data is fetched
-
   return (
     <div className="min-h-screen relative pl-16">
-      <div
-        className="absolute inset-y-0 left-0 right-0 bg-cover bg-center z-0 bg-[url('/image/bg.png')]"
-      />
+      <div className="absolute inset-y-0 left-0 right-0 bg-cover bg-center z-0 bg-[url('/image/bg.png')]" />
       <div className="relative z-10 max-w-4xl mx-auto pt-36 pb-12 px-6">
         <div className="bg-white/80 backdrop-blur-md rounded-3xl shadow-2xl p-8">
           <h1 className="text-3xl font-bold text-center text-green-800 mb-6">
-            Update Your Recipe
+            Add Your Recipe
           </h1>
+
           <div className="flex justify-between mb-8">
             {steps.map((label, index) => (
               <div
                 key={index}
-                className={`text-center flex-1 font-medium text-sm ${step === index
-                  ? "text-orange-600 border-b-2 border-orange-600"
-                  : "text-gray-400"
-                  }`}
+                className={`text-center flex-1 font-medium text-sm ${
+                  step === index
+                    ? "text-orange-600 border-b-2 border-orange-600"
+                    : "text-gray-400"
+                }`}
               >
                 {label}
               </div>
             ))}
           </div>
 
-          {/* Form section */}
           <form onSubmit={handleSubmit(onSubmit)} className="space-y-6">
             <motion.div
               key={step}
@@ -135,7 +111,6 @@ const UpdateRecipePage = () => {
               animate={{ opacity: 1, y: 0 }}
               transition={{ duration: 0.4 }}
             >
-              {/* Step 1: Basic Info */}
               {step === 0 && (
                 <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
                   <div>
@@ -143,7 +118,6 @@ const UpdateRecipePage = () => {
                     <input
                       {...register("title", { required: "Title is required" })}
                       className="w-full border rounded px-3 py-2"
-                      defaultValue={recipeData.title}
                     />
                     {errors.title && <p className="text-red-500 text-sm">{errors.title.message}</p>}
                   </div>
@@ -152,7 +126,6 @@ const UpdateRecipePage = () => {
                     <select
                       {...register("category", { required: "Category is required" })}
                       className="w-full border rounded px-3 py-2"
-                      defaultValue={recipeData.category}
                     >
                       <option value="">-- Select --</option>
                       <option>Main course</option>
@@ -164,85 +137,77 @@ const UpdateRecipePage = () => {
                   <div>
                     <label className="block font-medium">Cooking Time</label>
                     <input
-                      {...register("cookingTime", { required: "Cooking time is required" })}
+                      type="number"
+                      {...register("cookingTime", { required: "Required" })}
                       className="w-full border rounded px-3 py-2"
-                      defaultValue={recipeData.cookingTime}
                     />
                     {errors.cookingTime && <p className="text-red-500 text-sm">{errors.cookingTime.message}</p>}
                   </div>
                   <div>
                     <label className="block font-medium">Portion</label>
                     <input
-                      {...register("portion", { required: "Portion size is required" })}
+                      type="number"
+                      {...register("portion", { required: "Required" })}
                       className="w-full border rounded px-3 py-2"
-                      defaultValue={recipeData.portion}
                     />
                     {errors.portion && <p className="text-red-500 text-sm">{errors.portion.message}</p>}
                   </div>
                 </div>
               )}
 
-              {/* Step 2: Ingredients & Instructions */}
               {step === 1 && (
-                <div className="space-y-4">
+                <div className="space-y-3">
                   <div>
                     <label className="block font-medium">Ingredients</label>
                     <textarea
-                      {...register("ingredients", { required: "Ingredients are required" })}
-                      className="w-full border rounded px-3 py-2 h-32"
-                      defaultValue={recipeData.ingredients}
+                      {...register("ingredients", { required: "Required" })}
+                      className="w-full border rounded px-3 py-2 h-28"
                     />
                     {errors.ingredients && <p className="text-red-500 text-sm">{errors.ingredients.message}</p>}
                   </div>
                   <div>
                     <label className="block font-medium">Instructions</label>
                     <textarea
-                      {...register("instructions", { required: "Instructions are required" })}
-                      className="w-full border rounded px-3 py-2 h-40"
-                      defaultValue={recipeData.instructions}
+                      {...register("instructions", { required: "Required" })}
+                      className="w-full border rounded px-3 py-2 h-30"
                     />
                     {errors.instructions && <p className="text-red-500 text-sm">{errors.instructions.message}</p>}
                   </div>
                 </div>
               )}
 
-              {/* Step 3: Nutrition & Image */}
               {step === 2 && (
                 <div className="grid grid-cols-2 gap-4">
                   <div>
                     <label className="block font-medium">Calories</label>
                     <input
                       type="number"
-                      {...register("calories", { required: "Calories are required" })}
+                      {...register("calories", { required: "Required" })}
                       className="w-full border rounded px-3 py-2"
-                      defaultValue={recipeData.calories}
                     />
                   </div>
                   <div>
                     <label className="block font-medium">Protein</label>
                     <input
                       type="number"
-                      {...register("protein", { required: "Protein is required" })}
+                      {...register("protein", { required: "Required" })}
                       className="w-full border rounded px-3 py-2"
-                      defaultValue={recipeData.protein}
                     />
                   </div>
                   <div>
                     <label className="block font-medium">Fat</label>
                     <input
                       type="number"
-                      {...register("fat", { required: "Fat is required" })}
+                      {...register("fat", { required: "Required" })}
                       className="w-full border rounded px-3 py-2"
-                      defaultValue={recipeData.fat}
                     />
                   </div>
                   <div>
                     <label className="block font-medium">Carbs</label>
                     <input
                       type="number"
-                      {...register("carbs", { required: "Carbs are required" })}
+                      {...register("carbs", { required: "Required" })}
                       className="w-full border rounded px-3 py-2"
-                      defaultValue={recipeData.carbs}
                     />
                   </div>
                   <div className="col-span-2">
@@ -253,7 +218,9 @@ const UpdateRecipePage = () => {
                       {...register("imageFile")}
                       className="w-full border rounded px-3 py-2"
                     />
-                    {errors.imageFile && <p className="text-red-500 text-sm">{errors.imageFile.message}</p>}
+                    {errors.imageFile && (
+                      <p className="text-red-500 text-sm">{errors.imageFile.message}</p>
+                    )}
                   </div>
                 </div>
               )}
@@ -263,18 +230,15 @@ const UpdateRecipePage = () => {
               <button
                 type="button"
                 onClick={prevStep}
-                className={`px-6 py-2 border rounded text-gray-600 hover:bg-gray-100 ${step === 0 ? "invisible" : ""}`}
+                className={`px-6 py-2 border rounded text-gray-600 hover:bg-gray-100 ${step === 0 ? "invisible" : ""
+                  }`}
               >
                 Back
               </button>
-
               {step < steps.length - 1 ? (
                 <button
                   type="button"
-                  onClick={(event) => {
-                    event.preventDefault();
-                    nextStep();
-                  }}
+                  onClick={nextStep}
                   className="px-6 py-2 bg-green-600 text-white rounded hover:bg-green-700"
                 >
                   Next
@@ -284,7 +248,7 @@ const UpdateRecipePage = () => {
                   type="submit"
                   className="px-6 py-2 bg-orange-600 text-white rounded hover:bg-orange-700"
                 >
-                  Update
+                  Save
                 </button>
               )}
             </div>
@@ -295,4 +259,4 @@ const UpdateRecipePage = () => {
   );
 };
 
-export default UpdateRecipePage;
+export default CreateRecipePage;
