+++ src/app/RecipeManagement/ManageRecipe/UpdateRecipeForm/page.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { motion } from "framer-motion";
 import axios from "axios";
+import { useRouter, useSearchParams } from "next/navigation";
 import toast from 'react-hot-toast';
-import { useRouter } from 'next/navigation';
 
 type RecipeFormData = {
   title: string;
@@ -23,22 +23,44 @@ type RecipeFormData = {
 
 const steps = ["Basic Info", "Details", "Nutrition & Image"];
 
-const CreateRecipePage = () => {
-  const router = useRouter();
+const UpdateRecipePage = () => {
   const [step, setStep] = useState(0);
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-    reset,
-  } = useForm<RecipeFormData>();
+  const [isClient, setIsClient] = useState(false); // To track client-side rendering
+  const [recipeData, setRecipeData] = useState<RecipeFormData | null>(null);
+  const { register, handleSubmit, formState: { errors }, reset } = useForm<RecipeFormData>();
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const recipeId = searchParams.get('id'); // Get the recipe ID from search params
+  
+  // Set client-side flag after the component is mounted
+  useEffect(() => {
+    setIsClient(true);
+  }, []);
+
+  // Fetch recipe data when the page loads (client-side)
+  useEffect(() => {
+    if (recipeId && isClient) {
+      console.log("Fetching recipe with ID:", recipeId); // Log the recipeId to ensure it's correct
+      const fetchRecipeData = async () => {
+        try {
+          const response = await axios.get(`${process.env.NEXT_PUBLIC_API_BASE_URL}/api/Recipe/${recipeId}`);
+          console.log("Fetched recipe data:", response.data); // Log the fetched recipe data
+          setRecipeData(response.data);
+          reset(response.data); // Populate the form with fetched data
+        } catch (error) {
+          console.error("Error fetching recipe data:", error);
+          alert("Failed to load recipe data");
+        }
+      };
+      fetchRecipeData();
+    }
+  }, [recipeId, isClient, reset]);
 
   const nextStep = () => setStep((prev) => Math.min(prev + 1, steps.length - 1));
   const prevStep = () => setStep((prev) => Math.max(prev - 1, 0));
 
   const onSubmit = async (data: RecipeFormData) => {
-    console.log("Form submitted:", data);
-
+    console.log("Form updated:", data); // Log the form data on submit
     try {
       const formData = new FormData();
       formData.append("title", data.title);
@@ -56,23 +78,21 @@ const CreateRecipePage = () => {
         formData.append("image", data.imageFile[0]); // backend must accept this as IFormFile
       }
 
-      const response = await axios.post(`${process.env.NEXT_PUBLIC_API_BASE_URL}/api/Recipe/addRecipe1`, formData, {
-        headers: {
-          'Content-Type': 'multipart/form-data',
-        },
-      });
+      const response = await axios.put(`${process.env.NEXT_PUBLIC_API_BASE_URL}/api/Recipe/updateRecipe/${recipeId}`, formData);
+      console.log("Recipe updated successfully:", response.data);
+      toast.success("Recipe updated successfully!");
 
-      console.log("Recipe added successfully:", response.data);
-      toast.success("Recipe added successfully!");
       reset();
-      setStep(0);
-      setTimeout(() => {
-        router.replace('/Pages/RecipeManagement/ManageRecipe/ManageRecipe');
-      }, 500);
+      setStep(0); // Reset to the first step after submission
+
+        setTimeout(() => {
+          router.replace('/RecipeManagement/ManageRecipe/ManageRecipe'); // Navigate to the RecipeManage page
+        }, 500);
+      
     } catch (error: unknown) {
       if (axios.isAxiosError(error)) {
         console.error("Axios error:", error.response?.data || error.message);
-        toast.error("Failed to add recipe: " + (error.response?.data?.message || error.message));
+        toast.error("Failed to update recipe: " + (error.response?.data?.message || error.message));
       } else {
         console.error("Unexpected error:", error);
         toast.error("An unexpected error occurred.");
@@ -80,30 +100,34 @@ const CreateRecipePage = () => {
     }
   };
 
+  if (!isClient) return null; // Avoid rendering anything before client-side hydration
+  if (!recipeData) return <p>Loading recipe data...</p>; // Show loading message until recipe data is fetched
+
   return (
     <div className="min-h-screen relative pl-16">
-      <div className="absolute inset-y-0 left-0 right-0 bg-cover bg-center z-0 bg-[url('/image/bg.png')]" />
+      <div
+        className="absolute inset-y-0 left-0 right-0 bg-cover bg-center z-0 bg-[url('/image/bg.png')]"
+      />
       <div className="relative z-10 max-w-4xl mx-auto pt-36 pb-12 px-6">
         <div className="bg-white/80 backdrop-blur-md rounded-3xl shadow-2xl p-8">
           <h1 className="text-3xl font-bold text-center text-green-800 mb-6">
-            Add Your Recipe
+            Update Your Recipe
           </h1>
-
           <div className="flex justify-between mb-8">
             {steps.map((label, index) => (
               <div
                 key={index}
-                className={`text-center flex-1 font-medium text-sm ${
-                  step === index
-                    ? "text-orange-600 border-b-2 border-orange-600"
-                    : "text-gray-400"
-                }`}
+                className={`text-center flex-1 font-medium text-sm ${step === index
+                  ? "text-orange-600 border-b-2 border-orange-600"
+                  : "text-gray-400"
+                  }`}
               >
                 {label}
               </div>
             ))}
           </div>
 
+          {/* Form section */}
           <form onSubmit={handleSubmit(onSubmit)} className="space-y-6">
             <motion.div
               key={step}
@@ -111,6 +135,7 @@ const CreateRecipePage = () => {
               animate={{ opacity: 1, y: 0 }}
               transition={{ duration: 0.4 }}
             >
+              {/* Step 1: Basic Info */}
               {step === 0 && (
                 <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
                   <div>
@@ -118,6 +143,7 @@ const CreateRecipePage = () => {
                     <input
                       {...register("title", { required: "Title is required" })}
                       className="w-full border rounded px-3 py-2"
+                      defaultValue={recipeData.title}
                     />
                     {errors.title && <p className="text-red-500 text-sm">{errors.title.message}</p>}
                   </div>
@@ -126,6 +152,7 @@ const CreateRecipePage = () => {
                     <select
                       {...register("category", { required: "Category is required" })}
                       className="w-full border rounded px-3 py-2"
+                      defaultValue={recipeData.category}
                     >
                       <option value="">-- Select --</option>
                       <option>Main course</option>
@@ -137,77 +164,85 @@ const CreateRecipePage = () => {
                   <div>
                     <label className="block font-medium">Cooking Time</label>
                     <input
-                      type="number"
-                      {...register("cookingTime", { required: "Required" })}
+                      {...register("cookingTime", { required: "Cooking time is required" })}
                       className="w-full border rounded px-3 py-2"
+                      defaultValue={recipeData.cookingTime}
                     />
                     {errors.cookingTime && <p className="text-red-500 text-sm">{errors.cookingTime.message}</p>}
                   </div>
                   <div>
                     <label className="block font-medium">Portion</label>
                     <input
-                      type="number"
-                      {...register("portion", { required: "Required" })}
+                      {...register("portion", { required: "Portion size is required" })}
                       className="w-full border rounded px-3 py-2"
+                      defaultValue={recipeData.portion}
                     />
                     {errors.portion && <p className="text-red-500 text-sm">{errors.portion.message}</p>}
                   </div>
                 </div>
               )}
 
+              {/* Step 2: Ingredients & Instructions */}
               {step === 1 && (
-                <div className="space-y-3">
+                <div className="space-y-4">
                   <div>
                     <label className="block font-medium">Ingredients</label>
                     <textarea
-                      {...register("ingredients", { required: "Required" })}
-                      className="w-full border rounded px-3 py-2 h-28"
+                      {...register("ingredients", { required: "Ingredients are required" })}
+                      className="w-full border rounded px-3 py-2 h-32"
+                      defaultValue={recipeData.ingredients}
                     />
                     {errors.ingredients && <p className="text-red-500 text-sm">{errors.ingredients.message}</p>}
                   </div>
                   <div>
                     <label className="block font-medium">Instructions</label>
                     <textarea
-                      {...register("instructions", { required: "Required" })}
-                      className="w-full border rounded px-3 py-2 h-30"
+                      {...register("instructions", { required: "Instructions are required" })}
+                      className="w-full border rounded px-3 py-2 h-40"
+                      defaultValue={recipeData.instructions}
                     />
                     {errors.instructions && <p className="text-red-500 text-sm">{errors.instructions.message}</p>}
                   </div>
                 </div>
               )}
 
+              {/* Step 3: Nutrition & Image */}
               {step === 2 && (
                 <div className="grid grid-cols-2 gap-4">
                   <div>
                     <label className="block font-medium">Calories</label>
                     <input
                       type="number"
-                      {...register("calories", { required: "Required" })}
+                      {...register("calories", { required: "Calories are required" })}
                       className="w-full border rounded px-3 py-2"
+                      defaultValue={recipeData.calories}
                     />
                   </div>
                   <div>
                     <label className="block font-medium">Protein</label>
                     <input
                       type="number"
-                      {...register("protein", { required: "Required" })}
+                      {...register("protein", { required: "Protein is required" })}
                       className="w-full border rounded px-3 py-2"
+                      defaultValue={recipeData.protein}
                     />
                   </div>
                   <div>
                     <label className="block font-medium">Fat</label>
                     <input
                       type="number"
-                      {...register("fat", { required: "Required" })}
+                      {...register("fat", { required: "Fat is required" })}
                       className="w-full border rounded px-3 py-2"
+                      defaultValue={recipeData.fat}
                     />
                   </div>
                   <div>
                     <label className="block font-medium">Carbs</label>
                     <input
                       type="number"
-                      {...register("carbs", { required: "Required" })}
+                      {...register("carbs", { required: "Carbs are required" })}
                       className="w-full border rounded px-3 py-2"
+                      defaultValue={recipeData.carbs}
                     />
                   </div>
                   <div className="col-span-2">
@@ -218,9 +253,7 @@ const CreateRecipePage = () => {
                       {...register("imageFile")}
                       className="w-full border rounded px-3 py-2"
                     />
-                    {errors.imageFile && (
-                      <p className="text-red-500 text-sm">{errors.imageFile.message}</p>
-                    )}
+                    {errors.imageFile && <p className="text-red-500 text-sm">{errors.imageFile.message}</p>}
                   </div>
                 </div>
               )}
@@ -230,15 +263,18 @@ const CreateRecipePage = () => {
               <button
                 type="button"
                 onClick={prevStep}
-                className={`px-6 py-2 border rounded text-gray-600 hover:bg-gray-100 ${step === 0 ? "invisible" : ""
-                  }`}
+                className={`px-6 py-2 border rounded text-gray-600 hover:bg-gray-100 ${step === 0 ? "invisible" : ""}`}
               >
                 Back
               </button>
+
               {step < steps.length - 1 ? (
                 <button
                   type="button"
-                  onClick={nextStep}
+                  onClick={(event) => {
+                    event.preventDefault();
+                    nextStep();
+                  }}
                   className="px-6 py-2 bg-green-600 text-white rounded hover:bg-green-700"
                 >
                   Next
@@ -248,7 +284,7 @@ const CreateRecipePage = () => {
                   type="submit"
                   className="px-6 py-2 bg-orange-600 text-white rounded hover:bg-orange-700"
                 >
-                  Save
+                  Update
                 </button>
               )}
             </div>
@@ -259,4 +295,4 @@ const CreateRecipePage = () => {
   );
 };
 
-export default CreateRecipePage;
+export default UpdateRecipePage;
