+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import HomePage from "./Pages/RecipeManagement/Home/HomePage/page";
+import HomePage from "./RecipeManagement/Home/HomePage/page";
 
 export default function Home() {
   return (
