+++ src/components/SideBar.tsx
@@ -112,4 +112,4 @@ const Sidebar = ({ isOpen }: SidebarProps) => {
   );
 };
 
-export default Sidebar;
+export default Sidebar;A
