+++ src/components/SimpleFooter.tsx
